+++ derive/CHANGELOG.md
@@ -1,5 +1,6 @@
 Unreleased
 ----------
+- Added support for `gui(Event = ...)` attribute
 - Dropped "Gui" prefix from custom derive macros
 - Downgraded dependency to `gui` to a dev-dependency